pyproject.toml
[tool.poetry]
name = "tap_zenefits"
version = "0.1.0"
description = ""
authors = ["Jordan <jordan.wallace.williams@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
cchardet = "^2.1.6"
aiodns = "^2.0.0"
python-dotenv = "^0.14.0"
pyyaml = "^5.3.1"
vcrpy = "^4.1.0"
pytest-vcr = "^1.0.2"
pytest-asyncio = "^0.14.0"
coverage = "^5.3"
pytest-cov = "^2.10.1"
pandas = "^1.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
